src/requirements.txt
# SPDX-FileCopyrightText: Copyright (c) 2025 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
# http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.
#
# SPDX-License-Identifier: Apache-2.0

# NOTE: After modifying this file, locked_requirements.txt must be regenrated with the following
# command (Run in this directory)
# pip install pip-tools
# pip-compile ./requirements.txt -o locked_requirements.txt --allow-unsafe -v --generate-hashes

pydantic==1.10.13

# For job queue
kombu==5.2.4
redis==4.4.4

# kubernetes
kubernetes==24.2.0

# OSMO2 database
psycopg2-binary==2.9.10

# REST API
fastapi==0.115.5
uvicorn[standard]==0.30.1
h11==0.16.0

# JWT
pyjwt==2.10.1

# Requests
requests==2.32.4

# Jinja
jinja2==3.1.6

# S3
boto3==1.35.0
botocore==1.35.0
mypy-boto3-iam==1.35.*
mypy-boto3-s3==1.35.*
mypy-boto3-sts==1.35.*

cryptography==44.0.1

# JWE, JWK
jwcrypto==1.5.6

# Pyinstaller
pyinstaller==6.12.0

# Yaml
pyyaml==6.0.2
ijson==3.2.3

# Rendering on the client
texttable==1.6.4
tqdm==4.66.3
types-pytz==2023.3.0.0
pytz==2023.3

# Parsing
lark==1.1.5

# Opentelemetry Instrumentation
opentelemetry-exporter-otlp==1.20.0
opentelemetry-instrumentation-fastapi==0.41b0
opentelemetry-sdk==1.20.0

# Slack
slack_sdk==3.20.2

# Websocket App on the client
websocket-client==1.5.2
websockets==11.0.3

# CLI autocomplete package
shtab==1.6.4

wrapt==1.15.0

# MacOS build
macholib==1.16.3

# File system based cache
diskcache==5.6.3

# For tutorials
amqp==5.1.1

# Workspace Daemon
watchdog==6.0.0

# Azure Blob Storage
azure-storage-blob==12.26.0
azure-identity==1.23.1
